Added Command interface and classes and organized files into packages
