--- src/main/java/exceptions/EmptyTaskDescriptionException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class EmptyTaskDescriptionException extends Exception {
-    public EmptyTaskDescriptionException(String message) {
-        super(message);
-    }
-}
