--- src/main/java/exceptions/InvalidDateFormatException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class InvalidDateFormatException extends Exception {
-    public InvalidDateFormatException(String message) {
-        super(message);
-    }
-}
