--- src/main/java/exceptions/InvalidTaskNumberException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class InvalidTaskNumberException extends Exception {
-    public InvalidTaskNumberException(String message) {
-        super(message);
-    }
-}
