--- src/main/java/exceptions/TaskAlreadyDoneException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class TaskAlreadyDoneException extends Exception {
-    public TaskAlreadyDoneException(String message) {
-        super(message);
-    }
-}
