--- src/main/java/exceptions/TaskNotDoneException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class TaskNotDoneException extends Exception {
-    public TaskNotDoneException(String message) {
-        super(message);
-    }
-}
