+++ src/main/java/martin/commands/Command.java
@@ -1,0 +1,10 @@
+package martin.commands;
+
+import martin.exceptions.MartinException;
+
+public interface Command {
+    /**
+     * Executes the specific command.
+     */
+    void execute() throws MartinException;
+}
