+++ src/main/java/martin/exceptions/EmptyTaskDescriptionException.java
@@ -1,0 +1,7 @@
+package martin.exceptions;
+
+public class EmptyTaskDescriptionException extends MartinException {
+    public EmptyTaskDescriptionException(String message) {
+        super(message);
+    }
+}
