+++ src/main/java/martin/exceptions/InvalidCommandException.java
@@ -1,0 +1,7 @@
+package martin.exceptions;
+
+public class InvalidCommandException extends MartinException {
+    public InvalidCommandException(String message) {
+        super(message);
+    }
+}
