+++ src/main/java/martin/exceptions/InvalidDateFormatException.java
@@ -1,0 +1,7 @@
+package martin.exceptions;
+
+public class InvalidDateFormatException extends MartinException {
+    public InvalidDateFormatException(String message) {
+        super(message);
+    }
+}
