+++ src/main/java/martin/exceptions/InvalidTaskNumberException.java
@@ -1,0 +1,7 @@
+package martin.exceptions;
+
+public class InvalidTaskNumberException extends MartinException {
+    public InvalidTaskNumberException(String message) {
+        super(message);
+    }
+}
