+++ src/main/java/martin/exceptions/MartinException.java
@@ -1,0 +1,7 @@
+package martin.exceptions;
+
+public class MartinException extends Exception {
+    public MartinException(String message) {
+        super(message);
+    }
+}
