+++ src/main/java/martin/exceptions/TaskAlreadyDoneException.java
@@ -1,0 +1,7 @@
+package martin.exceptions;
+
+public class TaskAlreadyDoneException extends MartinException {
+    public TaskAlreadyDoneException(String message) {
+        super(message);
+    }
+}
