+++ src/main/java/martin/exceptions/TaskNotDoneException.java
@@ -1,0 +1,7 @@
+package martin.exceptions;
+
+public class TaskNotDoneException extends MartinException {
+    public TaskNotDoneException(String message) {
+        super(message);
+    }
+}
