+++ src/main/java/martin/parser/ListCommand.java
@@ -1,0 +1,5 @@
+package martin.parser;
+
+public class ListCommand {
+
+}
