+++ src/main/java/martin/task/Event.java
@@ -1,3 +1,5 @@
+package martin.task;
+
 public class Event extends Task {
     protected String from;
     protected String to;
